Favorite screen and handle on add/remove from fav click

+++ shared/src/commonMain/kotlin/az/zero/animeaz/data/local/database/mapper/Mappers.kt
@@ -4,21 +4,22 @@ import az.zero.animeaz.domain.model.Anime
 import database.AnimeEntity
 
 fun AnimeEntity.toAnime(): Anime {
-    // FIXME: the sql delight plugin should return int when the type is INT or INTEGER but instead it returns Long
+    // FIXME: the sql delight plugin should return int when the type is INT or
+    //  INTEGER but instead it returns Long so I will change Int to long for now 🙄🙄🙄🙄🙄🙄🙄🙄
     return Anime(
         id = id,
         name = name,
         image = image,
         cover = cover,
         score = score.toFloat(),
-        reviewCount = reviewCount.toInt(),
-        rank = rank.toInt(),
-        popularity = popularity.toInt(),
+        reviewCount = reviewCount,
+        rank = rank,
+        popularity = popularity,
         airingStatus = airingStatus,
         description = description,
         season = season,
         year = year,
-        numberOfEpisodes = numberOfEpisodes.toInt(),
+        numberOfEpisodes = numberOfEpisodes,
         showType = showType,
         genres = genres
     )
